+++ resources/js/components.js
@@ -16,7 +16,6 @@ import datePicker from './datePicker.js';
 import dialog from './dialog.js';
 import dropdownMenu from './dropdownMenu.js';
 import dropdownMenuSub from './dropdownMenuSub.js';
-import editor from './editor.js';
 import popover from './popover.js';
 import select from './select.js';
 import sheet from './sheet.js';
@@ -46,7 +45,6 @@ export {
     dialog,
     dropdownMenu,
     dropdownMenuSub,
-    editor,
     popover,
     select,
     sheet,
@@ -77,7 +75,6 @@ export default {
     dialog,
     dropdownMenu,
     dropdownMenuSub,
-    editor,
     popover,
     select,
     sheet,