+++ routes/assets.php
@@ -17,6 +17,18 @@ Route::prefix('april-ui')->name('april-ui.')->group(function () {
         ]);
     })->name('april.min.js');
 
+    Route::get('/editor.js', function () use ($distPath) {
+        return response()->file($distPath.'editor.js', [
+            'Content-Type' => 'text/javascript; charset=utf-8',
+        ]);
+    })->name('editor.js');
+
+    Route::get('/editor.min.js', function () use ($distPath) {
+        return response()->file($distPath.'editor.min.js', [
+            'Content-Type' => 'text/javascript; charset=utf-8',
+        ]);
+    })->name('editor.min.js');
+
     Route::get('/april.css', function () use ($distPath) {
         return response()->file($distPath.'april.css', [
             'Content-Type' => 'text/css; charset=utf-8',